// packages/pivo/src/open-api/open-api-types.ts
export declare namespace OpenAPIV3Semantics {
  interface Document {
    openapi: string
    info: InfoObject
    servers?: ServerObject[]
    paths: PathsObject
    components?: ComponentsObject
    security?: SecurityRequirementObject[]
    tags?: TagObject[]
    externalDocs?: ExternalDocumentationObject
    '@context'?: JsonLD.Context
  }
  interface InfoObject {
    title: string
    description?: string
    termsOfService?: string
    contact?: ContactObject
    license?: LicenseObject
    version: string
  }
  interface ContactObject {
    name?: string
    url?: string
    email?: string
  }
  interface LicenseObject {
    name: string
    url?: string
  }
  interface ServerObject {
    url: string
    description?: string
    variables?: {
      [variable: string]: ServerVariableObject
    }
  }
  interface ServerVariableObject {
    enum?: string[]
    default: string
    description?: string
  }
  interface PathsObject {
    [pattern: string]: PathItemObject
  }
  interface PathItemObject {
    $ref?: string
    summary?: string
    description?: string
    get?: OperationObject
    put?: OperationObject
    post?: OperationObject
    delete?: OperationObject
    options?: OperationObject
    head?: OperationObject
    patch?: OperationObject
    trace?: OperationObject
    servers?: ServerObject[]
    parameters?: Array<ReferenceObject | ParameterObject>
  }
  type OperationObject = {
    tags?: string[]
    summary?: string
    description?: string
    externalDocs?: ExternalDocumentationObject
    operationId?: string
    parameters?: Array<ReferenceObject | ParameterObject>
    requestBody?: ReferenceObject | RequestBodyObject
    responses?: ResponsesObject
    callbacks?: {
      [callback: string]: ReferenceObject | CallbackObject
    }
    deprecated?: boolean
    security?: SecurityRequirementObject[]
    servers?: ServerObject[]
  } & JsonLD.Entry
  interface ExternalDocumentationObject {
    description?: string
    url: string
  }
  interface ParameterObject extends ParameterBaseObject {
    name: string
    in: string
  }
  interface HeaderObject extends ParameterBaseObject {}
  type ParameterBaseObject = {
    description?: string
    required?: boolean
    deprecated?: boolean
    allowEmptyValue?: boolean
    style?: string
    explode?: boolean
    allowReserved?: boolean
    schema?: ReferenceObject | SchemaObject
    example?: any
    examples?: {
      [media: string]: ReferenceObject | ExampleObject
    }
    content?: {
      [media: string]: MediaTypeObject
    }
  } & JsonLD.Entry
  type NonArraySchemaObjectType =
    | 'null'
    | 'boolean'
    | 'object'
    | 'number'
    | 'string'
    | 'integer'
  type ArraySchemaObjectType = 'array'
  type SchemaObject = ArraySchemaObject | NonArraySchemaObject
  interface ArraySchemaObject extends BaseSchemaObject {
    type: ArraySchemaObjectType
    items: ReferenceObject | SchemaObject
  }
  interface NonArraySchemaObject extends BaseSchemaObject {
    type: NonArraySchemaObjectType
  }
// ...
  interface ReferenceObject {
    $ref: string
  }
// ...
  interface RequestBodyObject {
    description?: string
    content: ContentsObject
    required?: boolean
  }
  interface ResponsesObject {
    [code: string]: ReferenceObject | ResponseObject
  }
// ...
  interface CallbackObject {
    [url: string]: PathItemObject
  }
  interface SecurityRequirementObject {
    [name: string]: string[]
  }
  interface ComponentsObject {
    schemas?: {
      [key: string]: ReferenceObject | SchemaObject
    }
    responses?: {
      [key: string]: ReferenceObject | ResponseObject
    }
    parameters?: {
      [key: string]: ReferenceObject | ParameterObject
    }
    examples?: {
      [key: string]: ReferenceObject | ExampleObject
    }
    requestBodies?: {
      [key: string]: ReferenceObject | RequestBodyObject
    }
    headers?: {
      [key: string]: ReferenceObject | HeaderObject
    }
    securitySchemes?: {
      [key: string]: ReferenceObject | SecuritySchemeObject
    }
    links?: {
      [key: string]: ReferenceObject | LinkObject
    }
    callbacks?: {
      [key: string]: ReferenceObject | CallbackObject
    }
  }
// ...
  interface TagObject {
    name: string
    description?: string
    externalDocs?: ExternalDocumentationObject
  }
}

export declare namespace ExpandedOpenAPIV3Semantics {
// ...
  interface PathItemObject {
    $ref?: string
    summary?: string
    description?: string
    get?: OperationObject
    put?: OperationObject
    post?: OperationObject
    delete?: OperationObject
    options?: OperationObject
    head?: OperationObject
    patch?: OperationObject
    trace?: OperationObject
    servers?: ServerObject[]
    parameters?: Array<ParameterObject>
  }
// ...
}